cleaner code
