+++ config/connectDatabase.js
@@ -1,0 +1,14 @@
+const mongoose = require("mongoose");
+require("dotenv").config();
+const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.MONGO, {
+      useUnifiedTopology: true,
+      useNewUrlParser: true,
+    });
+  } catch (error) {
+    console.log(error);
+  }
+};
+
+module.exports = connectDB;
