+++ config/corsOptions.js
@@ -1,0 +1,4 @@
+const corsOptions = {
+  origin: "*",
+  optionsSuccessStatus: 200,
+};
