+++ index.js
@@ -1,22 +1,19 @@
+require("dotenv").config();
 const mongoose = require("mongoose");
 const express = require("express");
-const joi = require("joi");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-require("dotenv").config();
+const cors = require("cors");
+const corsOptions = require("./config/corsOptions");
+const connectDB = require("./config/connectDatabase");
 
-mongoose.connect(process.env.MONGO);
-
-const User = require("./models/user");
-const Paciente = require("./models/paciente");
-const Registro = require("./models/registro");
-const Psicologo = require("./models/psicologo");
+connectDB();
 
 const app = express();
 
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 3005;
 
 app.use(express.json());
+
+app.use(cors(corsOptions)); // <---
 
 const registro = require("./routers/registro");
 const auth = require("./routers/auth");
@@ -24,6 +21,7 @@ const auth = require("./routers/auth");
 app.use(registro);
 app.use(auth);
 
-app.listen(port, () => {
-    console.log(`Server running at ${port}`);
+mongoose.connection.once("open", () => {
+  console.log("Conectando ao db...");
+  app.listen(port, () => console.log(`Server running at ${port}`));
 });
