+++ package.json
@@ -19,6 +19,7 @@
     "homepage": "https://github.com/pedrovictor48/psicodaily-api#readme",
     "dependencies": {
         "bcrypt": "^5.1.0",
+        "cors": "^2.8.5",
         "dotenv": "^16.0.3",
         "express": "^4.18.2",
         "joi": "^17.9.1",
